requirements.txt
streamlit
opencv-python-headless
torch
torchvision
numpy
Pillow
roboflow
ultralytics
yolov5